Move to folder util and rename the bot_utils.js file
